zero_trust_gateway_phase_1/requirements.txt
# Zero Trust Gateway Phase 1 - Python Dependencies

# Database connectivity
psycopg2-binary==2.9.7

# Redis caching (optional)
redis==4.6.0

# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn==0.24.0

# JSON handling and utilities
python-json-logger==2.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Security and cryptography
cryptography==41.0.7

# Type checking and utilities
typing-extensions==4.8.0

# Environment variable management
python-dotenv==1.0.0

# Async database connections (optional upgrade)
asyncpg==0.29.0

# Logging enhancements
structlog==23.2.0 